feat: add delete method for role

+++ backend/model/role_model.go
@@ -10,7 +10,7 @@ type RoleCreateRequest struct {
 }
 
 type RoleUpdateRequest struct {
-	ID   uint16 `json:"id" validate:"required"`
+	ID   uint64 `json:"id" validate:"required"`
 	Name string `json:"name" validate:"required,max=100"`
 }
